wip; looking at how to return boolean from jooq delete
wip

add queries

build, but lots of wips

delete DatabaseConfigPersistence

pmd fix

build local -x test

fix names

port DatabaseConfigPersistenceTest tests to ActorDefinitionMigratorTest

clean up ActorDefinitionMigratorTest

add CRUD tests for standard sync

clean up ConfigRepository
